Create custom hook usePosts
